Extract delay line code to crate
